Do not rely on versioned paths when importing deps
This will remove the need to keep in sync the version in the psm1 with the
version in the .csproj.

Unfortunately, this is the best way I know of fixing issue #8.

There seems to be no automated way of downloading dependencies when
running Install-Module, so it's best to put DLLs inside the package itself.

Fixes #8