feat: use built-in Newtonsoft.Json if possible (#43)
If PowerShell ships with a version of Newtonsoft.Json, use it through an assembly binding.
If not, try to load the one we ship with the library.

This will reduce the likelihood of conflicts at assembly loading time, and makes the library
work under Powershell 6.0.x. Fixes #36.

Also, stop whenever there is an error importing the module, to make
debugging easier.